refactor(spatial): move upload_point_cloud out of NominalClient
Address PR feedback:
- Move upload_point_cloud to standalone function in nominal.core.spatial
- Remove upload_ouster_point_cloud (combined convert+upload); users
  should compose convert_ouster_dataset() + upload_point_cloud() instead
- Ouster thirdparty now only exports convert_ouster_dataset (render to disk)
- Export upload_point_cloud from nominal.core
